Users dashboard work updated

+++ app/Http/Controllers/RoleController.php
@@ -6,6 +6,7 @@ use App\Models\Role;
 use App\Models\Permission;
 use Illuminate\Http\Request;
 use App\Models\PermissionRole;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Controllers\RoleController;
 
 class RoleController extends Controller
@@ -25,7 +26,12 @@ class RoleController extends Controller
 
 }
 public function add()
-    {
+    {   
+        $permissionRole = PermissionRole::getPermission('Add Role',Auth::user()->role_id);
+        if(empty($permissionRole)){
+
+            abort(404);
+        }
         $getPermission = Permission::getRecord();
         $data['getPermission'] = $getPermission;
         return view('panel.role.add',$data);
@@ -35,7 +41,11 @@ public function insert(Request $request)
     {
 //    dd($request->all());
 
-    
+      $permissionRole = PermissionRole::getPermission(' Add Role',Auth::user()->role_id);
+if(empty($permissionRole)){
+
+    abort(404);
+}
        
         $save = new Role;
        
@@ -47,7 +57,12 @@ public function insert(Request $request)
     }
 
  public function edit($id)
-    {
+    {   
+        $permissionRole = PermissionRole::getPermission('Edit Role',Auth::user()->role_id);
+        if(empty($permissionRole)){
+
+            abort(404);
+        }
         $data['getRecord'] = Role::find($id);
        $data['getPermission'] = Permission::getRecord();
        $data['getRolePermission'] = PermissionRole::getRolePermission($id);
@@ -58,6 +73,11 @@ public function insert(Request $request)
 
     public function update($id,Request $request)
     {
+        $permissionRole = PermissionRole::getPermission('Edit Role',Auth::user()->role_id);
+        if(empty($permissionRole)){
+
+            abort(404);
+        }
         $save = Role::find($id);
         $save->name = $request->name;
         $save->save();
@@ -68,6 +88,11 @@ public function insert(Request $request)
 
     public function delete($id,Request $request)
     {
+        $permissionRole = PermissionRole::getPermission('Delete Role',Auth::user()->role_id);
+        if(empty($permissionRole)){
+
+            abort(404);
+        }
         $save = Role::find($id);
         $save->delete();
         return redirect('panel/role')->with('success','Role successfully deleted');
